refactor: RepositoryInitializer 제거
