+++ app/src/main/java/woowacourse/shopping/di/provider/RepositoryProvider.kt
@@ -4,7 +4,7 @@ import woowacourse.shopping.domain.repository.CartRepository
 import woowacourse.shopping.domain.repository.ProductRepository
 import woowacourse.shopping.domain.repository.RecentProductRepository
 
-object RepositoryProvider : RepositoryProviderInterface {
+object RepositoryProvider {
     lateinit var productRepository: ProductRepository
         private set
 
@@ -14,15 +14,15 @@ object RepositoryProvider : RepositoryProviderInterface {
     lateinit var recentProductRepository: RecentProductRepository
         private set
 
-    override fun initProductRepository(repository: ProductRepository) {
+    fun initProductRepository(repository: ProductRepository) {
         productRepository = repository
     }
 
-    override fun initCartRepository(repository: CartRepository) {
+    fun initCartRepository(repository: CartRepository) {
         cartRepository = repository
     }
 
-    override fun initRecentProductRepository(repository: RecentProductRepository) {
+    fun initRecentProductRepository(repository: RecentProductRepository) {
         recentProductRepository = repository
     }
 }
