--- app/src/main/java/woowacourse/shopping/di/provider/RepositoryProviderInterface.kt
@@ -1,13 +1,0 @@
-package woowacourse.shopping.di.provider
-
-import woowacourse.shopping.domain.repository.CartRepository
-import woowacourse.shopping.domain.repository.ProductRepository
-import woowacourse.shopping.domain.repository.RecentProductRepository
-
-interface RepositoryProviderInterface {
-    fun initProductRepository(repository: ProductRepository)
-
-    fun initCartRepository(repository: CartRepository)
-
-    fun initRecentProductRepository(repository: RecentProductRepository)
-}
